=htc Relocate HttpResponseParser.stage to tests
This stage is no longer required since 9dc474a1 converted HttpResponseParser
to a GraphStage.
